+++ back/app/models/entities.scala
@@ -12,7 +12,7 @@ case class Categories(catId: Long, tytul: String, opis: String)
 
 case class Products(prodId: Long, tytul: String, opis: String, catId: Long)
 
-case class ProductTypes(type_id: Long, title: String, price: Long, quantity: Long, product_id: Long)
+case class ProductTypes(type_id: Long, title: String, price: Float, quantity: Long, product_id: Long)
 
 //case class Purchases(purId: Long, prodId: Long, userId: Long)
